+++ dev/services/wms/ows_refactored/README.md
@@ -8,6 +8,7 @@ This file will act as an assembler which include all the layers required for the
 ### Resuable
 - legend definition
 - resource limit definition
+- custom functions
 
 ### Layers Groups
 ##### dev_only
